% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_emeta}
\alias{isobaric_emeta}
\title{Isobaric Peptide Feature Data (e_meta)}
\format{
A data frame with 215,220 rows (peptides) and 2 columns:
\describe{
  \item{Protein}{Protein identifier}
  \item{Peptide}{Peptide identifier}
}
}
\description{
A dataset containing the biomolecule metadata, a mapping of peptides to proteins.
}
