% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_fdata}
\alias{isobaric_fdata}
\title{Isobaric Peptide Feature Data (f_data)}
\format{
A data.frame with 12 rows (samples) and 5 columns (sample identifier,
  experiment/plate, channel, reference information, and group information):
\describe{
  \item{Sample}{Sample identifier (matches column headers in isobaric_edata)}
  \item{Set}{iTRAQ plate/experiment information}
  \item{iTRAQ.Channel}{Information on which channel a sample was run on}
  \item{Reference}{Indicator of whether a sample is a reference sample or not}
  \item{Group}{Group/treatment information}
}
}
\description{
A dataset containing the sample metadata.
}
