% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_object}
\alias{isobaric_object}
\title{Isobaric Data Object of Class isobaricpepData}
\format{
A isobaricpepData object (see
  \code{\link[pmartR]{as.isobaricpepData}} for details)
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data frame of expression data, where
  \eqn{p} is the number of peptides observed and \eqn{n} is the number of
  samples. Each row corresponds to data for each peptide}
  \item{f_data}{a data frame with \eqn{n} rows. Each row corresponds to a
  sample with one column giving the unique sample identifiers found in e_data
  column names and other columns providing qualitative and/or quantitative
  traits of each sample.}
  \item{e_meta}{a data frame containing biomolecule metadata}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data), and consist of a subset of de-identified data from a
multi-omic study . These data consist of samples from three different strains
of a virus, plus pooled reference samples. Since this is only a subset of
data from this study, the number of samples is not identical across TMT
plexes, and the TMT plex names are not sequential.
}
