% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_edata_neg}
\alias{lipid_edata_neg}
\title{Negative Ion Lipidomics Expression Data (e_data)}
\format{
A data.frame with 143 rows (lipids) and 12 columns (lipid identifier
  and samples):
\describe{
  \item{LipidCommonName}{Lipid common name}
  \item{Mock*}{Three columns of mock infected samples}
  \item{Infection*}{Eight columns of infected samples}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
  Data is also available on MassIVE (massive.ucsd.edu) with dataset id
  MSV000079154.
}
\description{
A dataset containing the quantified mass spectra for 143 lipids collected on
a Thermo LTQ-Orbitrap Velos Mass Spectrometer (Thermo Fisher Corporation, San
Jose, CA, USA) in negative ion mode.
}
