% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_fdata_neg}
\alias{lipid_fdata_neg}
\title{Negative Ion Lipidomics Sample Feature Data (f_data)}
\format{
A data.frame with 11 rows (samples) and 2 columns (sample identifier
  and condition):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in lipid_edata)}
  \item{Condition}{Character string indicating either "Infection" or "Mock"}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
  Data is also available on MassIVE (massive.ucsd.edu) with dataset id
  MSV000079154.
}
\description{
A dataset containing the sample metadata.
}
