% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_neg_emeta}
\alias{lipid_neg_emeta}
\title{Negative Ion Lipidomics Biomolecule Metadata (e_meta)}
\format{
A data frame with 337 rows (peptides) and 3 columns:
\describe{
  \item{Lipid}{Lipid identifier}
  \item{Row}{Lipid row}
  \item{Retention_Time}{Lipid retention time}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data). A dataset containing biomolecule metadata.
}
