% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_neg_fdata}
\alias{lipid_neg_fdata}
\title{Negative Ion Lipidomics Sample Feature Data (f_data)}
\format{
A data.frame with 11 rows (samples) and 2 columns (sample identifier
  and condition):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in e_data)}
  \item{Virus}{Strain of virus for each sample}
  \item{Donor}{Which donor the sample originated from}
  \item{Replicate}{Biological replicate number}
}
}
\description{
A dataset containing the sample metadata.
}
