% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_pos_edata}
\alias{lipid_pos_edata}
\title{Positive Ion Lipidomics Expression Data (e_data)}
\format{
A data.frame with 275 rows (lipids) and 46 columns (lipid identifier
  and samples):
\describe{
  \item{Lipid}{Lipid common name}
  \item{StrainA*}{Columns for Strain A samples}
  \item{StrainB*}{Columns for Strain B samples}
  \item{StrainC*}{Columns for Strain C samples}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data). A dataset containing a subset of de-identified data from a
multi-omic study (labeled proteomics, lipidomics, and RNAseq data). This data
has samples from three different strains of a virus.
}
