% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metab_edata}
\alias{metab_edata}
\title{Metabolomics Expression Data (e_data)}
\format{
A data.frame with 250 rows (metabolites) and 51 columns (metabolite
  identifier and samples):
\describe{
  \item{Metabolite}{Metabolite name}
  \item{Sample_*_Phenotype1_*}{Columns for samples belonging to the Phenotype
  1 group}
  \item{Sample_*_Phenotype2_*}{Columns for samples belonging to the Phenotype
  2 group}
  \item{Sample_*_Phenotype3_*}{Columns for samples belonging to the Phenotype
  3 group}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data). A dataset containing the quantified mass spectra
for 250 metabolites across 50 samples, collected via GC-MS.
}
