% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metab_object}
\alias{metab_object}
\title{Metabolomics Data Object of Class metabData}
\format{A metabData object:
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of metabolites observed and \eqn{n} is the number of samples. Each row corresponds to data for each metabolite}
  \item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}
  \item{e_meta}{NULL}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers. Data is also available on MassIVE (massive.ucsd.edu) with dataset id MSV000079153.
}
\description{
An S3 object of class metabData (see \code{\link[pmartR]{as.metabData}} for details)
}
