% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nmr_identified_edata}
\alias{nmr_identified_edata}
\title{NMR (Nuclear Magnetic Resonance) Expression Data (e_data)}
\format{
A data.frame with 38 rows (metabolites) and 42 columns (metabolite
  identifier and samples):
\describe{
  \item{Metabolite}{Metabolite name}
  \item{F*}{Sample names}
}
}
\description{
A dataset containing the NMR abundances for 38 metabolites in 41 samples 
taken from tomato plants.
}
\references{
Biais B, Benard C, Beauvoit B, Colombie S, Prodhomme D, Menard G, 
Bernillon S, Gehl B, Gautier H, Ballias P, et al. (2014), \emph{Remarkable 
reproducibility of enzyme activity profiles in tomato fruits grown under 
contrasting environments provides a roadmap for studies of fruit metabolism}.
Plant Physiol 164 (3), 1204-1221. doi: 10.1104/pp.113.231241.
}
