% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nmr_identified_emeta}
\alias{nmr_identified_emeta}
\title{NMR (Nuclear Magnetic Resonance) Biomolecule Metadata (e_meta)}
\format{
A data frame with 38 rows (metabolites) and 2 columns:
\describe{
  \item{Metabolite}{Metabolite identifier}
  \item{Bin}{Bin}
}
}
\description{
A dataset containing biomolecule metadata.
}
\references{
Biais B, Benard C, Beauvoit B, Colombie S, Prodhomme D, Menard G, 
Bernillon S, Gehl B, Gautier H, Ballias P, et al. (2014), \emph{Remarkable 
reproducibility of enzyme activity profiles in tomato fruits grown under 
contrasting environments provides a roadmap for studies of fruit metabolism}.
Plant Physiol 164 (3), 1204-1221. doi: 10.1104/pp.113.231241.
}
