% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nmr_identified_fdata}
\alias{nmr_identified_fdata}
\title{NMR (Nuclear Magnetic Resonance) Sample Metadata (f_data)}
\format{
A data.frame with 41 rows (samples) and 5 columns (sample identifier
  and condition):
\describe{
  \item{SampleID}{Sample identifier that matches column headers in pep_edata}
  \item{Condition}{Character string indicating either "Control" or "Shadow"}
  \item{Stage}{Character string denoting stage of the experiment, or time}
  \item{Time}{Numeric value for time in days at which the sample was
  harvested}
  \item{Concentration}{Sample concentration}
}
}
\description{
A dataset containing the sample metadata.
}
\references{
Biais B, Benard C, Beauvoit B, Colombie S, Prodhomme D, Menard G, 
Bernillon S, Gehl B, Gautier H, Ballias P, et al. (2014), \emph{Remarkable 
reproducibility of enzyme activity profiles in tomato fruits grown under 
contrasting environments provides a roadmap for studies of fruit metabolism}.
Plant Physiol 164 (3), 1204-1221. doi: 10.1104/pp.113.231241.
}
