% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nmr_identified_object}
\alias{nmr_identified_object}
\title{NMR (Nuclear Magnetic Resonance) Data Object of Class nmrData}
\format{
A nmrData object
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where
  \eqn{p} is the number of metabolites observed and \eqn{n} is the number of
  samples. Each row corresponds to abundance data for each metabolite.}
  \item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a
  sample with one column giving the unique sample identifiers found in e_data
  column names and other columns providing qualitative and/or quantitative
  traits of each sample.}
\item{e_meta}{a data.frame with \eqn{n} rows. Each row corresponds to a
  metabolite, with one column giving the unique metabolite identifier that is
  found in e_data, and additional columns providing more information for each
  metabolite.} }
}
\description{
An S3 object of class nmrData (see \code{\link[pmartR]{as.nmrData}} for
details)
}
\references{
Biais B, Benard C, Beauvoit B, Colombie S, Prodhomme D, Menard G, 
Bernillon S, Gehl B, Gautier H, Ballias P, et al. (2014), \emph{Remarkable 
reproducibility of enzyme activity profiles in tomato fruits grown under 
contrasting environments provides a roadmap for studies of fruit metabolism}.
Plant Physiol 164 (3), 1204-1221. doi: 10.1104/pp.113.231241.
}
