% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pep_edata}
\alias{pep_edata}
\title{Peptide-level Expression Data (e_data)}
\format{
A data.frame with 21,329 rows (peptides) and 25 columns (peptide
  identifier and samples):
\describe{
  \item{Peptide}{Unique character string for peptide identifier}
  \item{Sample_*_Phenotype1_*}{Columns for samples belonging to the Phenotype
  1 group}
  \item{Sample_*_Phenotype2_*}{Columns for samples belonging to the Phenotype
  2 group}
  \item{Sample_*_Phenotype3_*}{Columns for samples belonging to the Phenotype
  3 group}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data).
}
