% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pep_emeta}
\alias{pep_emeta}
\title{Peptide-level Meta Data (e_meta)}
\format{A data.frame with 17,407 rows (peptides) and 4 columns (peptide identifiers and meta data):
\describe{
  \item{Mass_Tag_ID}{Peptide mass tag identifier (matches Mass_Tag_ID in pep_edata)}
  \item{Protein}{Protein identifier}
  \item{Ref_ID}{Protein reference identifier number}
  \item{Peptide_Sequence}{Peptide sequence}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers. Data is also available on MassIVE (massive.ucsd.edu) with dataset id MSV000079152.
}
\description{
A dataset containing the peptide metadata (e.g. mapping from peptide to proteins).
}

