% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pep_fdata}
\alias{pep_fdata}
\title{Peptide-level Sample Feature Data (f_data)}
\format{A data.frame with 12 rows (samples) and 2 columns (sample identifier and condition):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in pep_edata)}
  \item{Condition}{Character string indicating either "Infection" or "Mock"}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers. Data is also available on MassIVE (massive.ucsd.edu) with dataset id MSV000079152.
}
\description{
A dataset containing the sample metadata.
}

