% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pep_techrep_fdata}
\alias{pep_techrep_fdata}
\title{Peptide-level Technical Replicate Sample Metadata (f_data)}
\format{
A data.frame with 64 rows (samples) and 4 columns:
\describe{
  \item{RunID}{LC-MS run identifier (matches column headers in pep_edata)}
  \item{FACTOR}{Character string indicating either regular weight (RW), or
  obese (OB) groups}
  \item{DILUTION}{Character string indicating dilution of mouse plasma to
  Shewanella Oneidensis MR-1}
  \item{TECH_REP}{Character string indicating which technical replicates
  belong to the same biological sample}
}
}
\description{
A dataset containing the technical replicate metadata, including technical
replicate sample assignment variable.
}
\references{
Webb-Robertson BJ, Matzke MM, Datta S, Payne SH, Kang J, Bramer
  LM, Nicora CD, Shukla AK, Metz TO, Rodland KD, Smith RD (2014),
  \emph{Bayesian proteoform modeling improves protein quantification of
  global proteomic measurements}. Molecular & cellular proteomics 13 (12),
  3639-46.
}
