% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pro_edata}
\alias{pro_edata}
\title{Protein-level Expression Data (e_data)}
\format{
A data.frame with 2,982 rows (proteins) and 25 columns (protein
  identifier and samples):
\describe{
  \item{RazorProtein}{Protein identifier}
  \item{Sample_*_Phenotype1_*}{Columns for samples belonging to the Phenotype
  1 group}
  \item{Sample_*_Phenotype2_*}{Columns for samples belonging to the Phenotype
  2 group}
  \item{Sample_*_Phenotype3_*}{Columns for samples belonging to the Phenotype
  3 group}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data). A dataset containing the log2 median-normalized
quantified mass spectra for 2,982 proteins from 24 samples.
}
