% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pro_emeta}
\alias{pro_emeta}
\title{Protein-level Biomolecule Metadata (e_meta)}
\format{
A data.frame with 2,982 rows (proteins) and 3 columns (protein
  identifier and meta data):
\describe{
  \item{RazorProtein}{Protein identifier that matches RazorProtein column in
  pep_edata}
  \item{peps_per_pro}{Integer count of number of peptides for each protein; 
  automatically calculated in protein_quant function in pmartR}
  \item{n_peps_used}{Integer count of the number of peptides used when
  rolling up to the protein level; automatically calculated in protein_quant
  function in pmartR}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data). A dataset containing the protein metadata 
(resulting from the roll-up of peptides to proteins within pmartR).
}
