% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pro_fdata}
\alias{pro_fdata}
\title{Protein-level Sample Metadata (f_data)}
\format{
A data.frame with 24 rows (samples) and 3 columns (sample identifier
  and condition):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in pro_edata)}
  \item{Phenotype}{Character string indicating Phenotype 1, 2, or 3}
  \item{SecondPhenotype}{Character string indicating A or B}
  \item{Characteristic}{Numeric value corresponding to a measured
  characteristic for each sample}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data). A dataset containing the sample metadata.
}
