% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rnaseq_emeta}
\alias{rnaseq_emeta}
\title{RNAseq Feature Data (e_meta)}
\format{
A data frame with 53,908 rows (transcript-by-gene combinations) and 2 columns:
\describe{
  \item{Transcript}{Transcript name}
  \item{Gene}{Gene name}
}
}
\description{
A dataset containing biomolecule metadata.
}
