% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rnaseq_fdata}
\alias{rnaseq_fdata}
\title{RNAseq Sample Metadata (f_data)}
\format{
A data.frame with 45 rows (samples) and 4 columns (sample identifier
  and condition):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in e_data)}
  \item{Virus}{Strain of virus for each sample}
  \item{Donor}{Which donor the sample originated from}
  \item{Replicate}{Biological replicate number}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data). A dataset containing the sample metadata.
}
