% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smoke_edata}
\alias{smoke_edata}
\title{Peptide-level Expression Data (e_data) from a cancer/obesity study on mice.}
\format{
A data.frame with 12328 rows (mass tags) and 17 columns (mass tag and sample columns):
\describe{
  \item{Mass_Tag_ID}{Peptide mass tag identifier}
  \item{U54_SMOKEmp_1*}{Eight columns of control samples}
  \item{U54_SMOKEmp_3*}{Eight columns of smoke exposed samples}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
A dataset containing the quantified mass spectra for 5,244 unique mass tag identifiers and 3646 unique proteins from a study examining the effect of inhalation endotoxin exposure and obesity on lung inflammation in mice.  This is a subset taken from the full data.
}
\references{
Tilton, S. C.; Waters, K. M.; Karin, N. J.; Webb-Robertson, B. J.; Zangar, R. C.; Lee, K. M.; Bigelow, D. J.; Pounds, J. G.; Corley, R. A., \emph{Diet-induced obesity reprograms the inflammatory response of the murine lung to inhaled endotoxin}. Toxicol Appl Pharmacol 2013, 267 (2), 137-48.
}
