% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smoke_emeta}
\alias{smoke_emeta}
\title{Peptide-level Meta Data (e_meta) from a cancer/obesity study on mice.}
\format{
A data.frame with 12328 rows (mass tags) and 6 columns:
\describe{
  \item{PeptideEx}{Raw peptide sequence strings.}
  \item{Mass_Tag_ID}{Indicator of control or smoke exposed group}
  \item{Peptide}{Unique peptide sequences}
  \item{Protein}{Mapped protein identifiers}
  \item{Mod_Count/Mod Description}{Number and descriptions of modifications}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
A dataset containing the peptide metadata (e.g. mapping from peptide to proteins) from a study examining the effect of inhalation endotoxin exposure and obesity on lung inflammation in mice.  This is a subset taken from the full data.
}
\references{
Tilton, S. C.; Waters, K. M.; Karin, N. J.; Webb-Robertson, B. J.; Zangar, R. C.; Lee, K. M.; Bigelow, D. J.; Pounds, J. G.; Corley, R. A., \emph{Diet-induced obesity reprograms the inflammatory response of the murine lung to inhaled endotoxin}. Toxicol Appl Pharmacol 2013, 267 (2), 137-48.
}
