% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smoke_fdata}
\alias{smoke_fdata}
\title{Peptide-level Sample Feature Data (f_data) from a cancer/obesity study on mice.}
\format{
A data.frame with 16 rows (samples) and 2 columns (sample names and group assignment):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in pep_edata)}
  \item{Status}{Indicator of control or smoke exposed group}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
A dataset containing the sample metadata from a study examining the effect of inhalation endotoxin exposure and obesity on lung inflammation in mice.  This is a subset taken from the full data.
}
\references{
Tilton, S. C.; Waters, K. M.; Karin, N. J.; Webb-Robertson, B. J.; Zangar, R. C.; Lee, K. M.; Bigelow, D. J.; Pounds, J. G.; Corley, R. A., \emph{Diet-induced obesity reprograms the inflammatory response of the murine lung to inhaled endotoxin}. Toxicol Appl Pharmacol 2013, 267 (2), 137-48.
}
