% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{techrep_pep_object}
\alias{techrep_pep_object}
\title{Peptide-level Data Object of Class pepData with technical replicates}
\format{A pepData object with technical replicate column "TECH_REP" found in f_data.
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of peptides observed and \eqn{n} is the number of technical replicates. Each row corresponds to data for each peptide}
  \item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a technical replicate with one column giving the unique replicate identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each replicate.\cr
  In this data.frame, the column "TECH_REP" specifies which technical replicates were drawn from the same biological sample.}
  \item{e_meta}{NULL}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
An S3 object of class pepData with a technical replicate column specified.  See documentation in \code{as.pepData} and \code{combine_techreps}.
}
