% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_edata}
\alias{isobaric_edata}
\title{Isobaric Peptide Expression Data (e_data)}
\format{A data.frame with 1993 rows (peptides) and 13 columns (peptide identifier and samples):}
\description{
A dataset containing simulated data from an iTRAQ instrument.
}
