% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_emeta}
\alias{isobaric_emeta}
\title{Isobaric Peptide Biomolecule Metadata (e_meta)}
\format{
A data frame with 215,220 rows (peptides) and 2 columns:
\describe{
  \item{Protein}{Protein identifier}
  \item{Peptide}{Peptide identifier}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data), and consist of a subset of de-identified data from a
multi-omic study. This dataset contains the biomolecule metadata, a mapping
of peptides to proteins.
}
