% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isobaric_fdata}
\alias{isobaric_fdata}
\title{Isobaric Peptide Feature Data (f_data)}
\format{
A data frame with 50 rows (samples) and 6 columns (sample identifier,
  experiment/plate information, and information about each sample):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in isobaric_edata)}
  \item{Plex}{iTRAQ plate/experiment information}
  \item{Virus}{Strain of virus for each sample}
  \item{Donor}{Which donor the sample originated from}
  \item{Replicate}{Biological replicate number}
  \item{TMT_day_plex_order}{The day, plex, and order on plate for the sample}
}
}
\description{
A dataset containing the sample metadata.
}
