% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_neg_emeta}
\alias{lipid_neg_emeta}
\title{Negative Ion Lipidomics Feature Data (e_meta)}
\format{
A data frame with 275 rows (peptides) and 3 columns:
\describe{
  \item{Lipid}{Lipid identifier}
  \item{Row}{Lipid row}
  \item{Retention_Time}{Lipid retention time}
}
}
\description{
A dataset containing biomolecule metadata.
}
