% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lipid_neg_fdata}
\alias{lipid_neg_fdata}
\title{Negative Ion Lipidomics Sample Metadata (f_data)}
\format{
A data.frame with 45 rows (samples) and 4 columns (sample identifier
  and associated information):
\describe{
  \item{SampleID}{Sample identifier (matches column headers in e_data)}
  \item{Virus}{Strain of virus for each sample}
  \item{Replicate}{Biological replicate number}
  \item{Donor}{Donor from which the sample originated}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data). A dataset containing the sample metadata.
}
