% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{metab_edata}
\alias{metab_edata}
\title{Metabolomics Expression Data (e_data)}
\format{
A data.frame with 80 rows (metabolites) and 13 columns (metabolite identifier and samples):
\describe{
  \item{Metabolite}{Metabolite name}
  \item{Mock*}{Three columns of mock infected samples}
  \item{Infection*}{Nine columns of infected samples}
}
}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers. Data is also available on MassIVE (massive.ucsd.edu) with dataset id MSV000079153.
}
\description{
A dataset containing the quantified mass spectra for 80 metabolites collected on an Agilent GC 7890A coupled with MSD 5975C Mass Spectrometer (Agilent Technologies, Santa Clara, CA, USA).
}
