% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pep_emeta}
\alias{pep_emeta}
\title{Peptide-level Biomolecule Metadata (e_meta)}
\format{
A data.frame with 21,329 rows (peptides) and 3 columns (peptide
  identifiers and meta data):
\describe{
  \item{Peptide}{Peptide identifier that matches Peptide column in
  pep_edata}
  \item{ProteinList}{Character string containing one or more proteins,
  separated by semi-colons}
  \item{RazorProtein}{Character string specifying the single protein which is
  the best mapping for the peptide; this should be used for protein roll-up}
}
}
\description{
These data are part of Experiment 2 (metabolomics, unlabeled peptide and
associated protein data). A dataset containing the peptide metadata (e.g.
mapping from peptide to proteins).
}
