% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pro_edata}
\alias{pro_edata}
\title{Protein-level Expression Data (e_data)}
\format{A data.frame with 2,731 rows (proteins) and 12 columns (protein identifier and samples):
\describe{
  \item{Reference}{Reference protein name}
  \item{Mock*}{Three columns of mock samples}
  \item{Infection*}{Eight columns of infection samples}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
A dataset containing the log2 median-normalized quantified mass spectra for 2,731 proteins (quantified from \code{\link{pep_edata}}) using Bayesian proteoform modeling (Webb-Robertson, 2014).
}
\references{
Webb-Robertson BJ, Matzke MM, Datta S, Payne SH, Kang J, Bramer LM, Nicora CD, Shukla AK, Metz TO, Rodland KD, Smith RD, Tardiff MF, McDermott JE, Pounds JG, Waters KM (2014), \emph{Bayesian proteoform modeling improves protein quantification of global proteomic measurements}. Molecular & Cellular Proteomics. doi: 10.1074/mcp.M113.030932.
}

