% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pro_object}
\alias{pro_object}
\title{Protein-level Data Object of Class proData}
\format{A proData object
\describe{
  \item{e_data}{a \eqn{p \times n + 1} data.frame of expression data, where \eqn{p} is the number of proteins observed and \eqn{n} is the number of samples. Each row corresponds to log2 median-normalized data for each protein.}
  \item{f_data}{a data.frame with \eqn{n} rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.}
  \item{e_meta}{NULL}
}}
\source{
See details of \code{\link{pmartRdata}} for relevant grant numbers.
}
\description{
An S3 object of class proData (see \code{\link[pmartRqc]{as.proData}} for details)
}
