% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rnaseq_edata}
\alias{rnaseq_edata}
\title{RNAseq Expression Data (e_data)}
\format{
A data.frame with 49,568 rows (transcripts) and 46 columns (transcript identifier
  and samples) containing transcript counts:
\describe{
  \item{Transcript}{Transcript name}
  \item{StrainA*}{Columns for Strain A samples}
  \item{StrainB*}{Columns for Strain B samples}
  \item{StrainC*}{Columns for Strain C samples}
}
}
\description{
A dataset containing a subset of de-identified data from a multi-omic study
(labeled proteomics, lipidomics, and RNAseq data). This data has samples from
three different strains of a virus.
}
