% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rnaseq_emeta}
\alias{rnaseq_emeta}
\title{RNAseq Biomolecule Metadata (e_meta)}
\format{
A data frame with 53,908 rows (transcript-by-gene combinations) and 2
  columns:
\describe{
  \item{Transcript}{Transcript name}
  \item{Gene}{Gene name}
}
}
\description{
These data are part of Experiment 1 (isobaric labeled proteomics, lipidomics,
and RNAseq data). A dataset containing biomolecule metadata.
}
